+++ CHANGELOG.md
@@ -6,5 +6,8 @@ rename `Unreleased` topic with the new version tag. Finally, create a new `Unrel
 
 ## Unreleased
 
+- Renamed `shouldRunInBackground` option to `background`.
+- Supports `stdIn` as an optional input.
+- Returns multiple values: `stdOut`, `stdErr`, `pid`, and `exitCode`.